+++ Brewview/src/main/java/com/repository/BeverageRepository.java
@@ -4,6 +4,7 @@ import com.model.Beverage;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 import java.util.List;
 import java.util.UUID;
@@ -15,9 +16,9 @@ public interface BeverageRepository extends JpaRepository<Beverage, UUID>{
 
     List<Beverage> findByType(String type);
 
-    List<Beverage> findByAbvBetween(int minAbv, int maxAbv);
+    List<Beverage> findByAbvBetween(BigDecimal minAbv, BigDecimal maxAbv);
 
-    List<Beverage> findByTypeAndAbvBetween(String type, int minAbv, int maxAbv);
+    List<Beverage> findByTypeAndAbvBetween(String type, BigDecimal minAbv, BigDecimal maxAbv);
 
     boolean existsByBeverageName(String beverageName);
